[staking] fix change candidate bug (#2572) (#2572)
When the new candidate is same as old one, staking handler are dealing
with 2 copies of the same candidate: vote of the first being deducted,
and vote of the second being added -- yet the second change overwrites
the first, b/c these 2 are the same candidate. This caused the total
vote to be incorrectly increased.